More work on cleaning up the API
